backend/account: fix race conditions between Initalize() and Close()
An account could be Initialized() after it was already Closed() if
they run at the same time. In ETH, the isClosed check was missing
altogether, but even in btc, there was a race condition. We need to
use the same lock for both.

The `WaitSynchronized()` in the Close function of the ETH account also
seems unneccesary and if it blocks the Close indefinitely, the DB
cannot be closed and the an account with the same DB cannot be
initialized (e.g. in `ReinitializeAccounts()`). Not sure why it was
added, the commit that introduced it had no explanation.

I ran into these issues while working on unit tests for accounts sync.