frontend: fix default & active currencies selection
Previously, we could:
1. set a default currency that's
not even in our active currencies list.

2. remove our default currency from our
active currencies list. So our default currency
won't be among the active currencies, which doesn't make sense.

These are bugs and aren't the intended behaviours.

The root cause is because we needed to add the checks.
After adding the checks though, it's still not fixed because we can't
get the updated values of the store once these values change,
except by refreshing it (which isn't what we want).

To fix this, we unfortunately still need to use our "old"
`share` decorator. It's because these values are currently
stored within our own `store` implementation. We can't
listen to changes in this store simply by using `useEffect`
since it's not a react state.

Thus, we need to wrap the component, in this case `activeCurrenciesDropdown`
using the `share` decorator to be able to subscribe to these
currencies changes and get the update immediately once the `store` values
get updated.