backend: mock btc/eth accounts in unit tests
By being able to mock the accounts, it is much easier to write unit
tests that rely on specific acocunt behavior. For example, in accounts
discovery we check if `.Transactions()` returns an empty or non empty
list to decide when to stop scanning accounts. Without mocking the
account, we'd have to connect to mocked Electrum/Etherscan servers,
which is much harder to do and maintain.