backend/handlers: use handlers that can't return an error
The error returned by handlers is not propagated to the frontend in
production. For errors that should be displayed to the user, we
currently use the `{ success: false }` pattern, sometimes with an
error message and/or error code.

It has been a frequent confusion that one can simply return an error
from the handler. This commit adds a non-error version of the handler
function, with the goal that all handlers eventually use this one, so
errors are properly handled in all cases.