frontend: add checklist
Moved the list of security notices, that have to each be checked
when creating a new wallet, to its own component. The state of each
checkbox does not have to be in the parent (bitbox02.tsx) as it is
only used to disable the continue button inside this view.

Note there are 2 similar list with security notices, but without
checkboxes, that are currently used as restore success view. It
probably makes sense to move the other 2 lists into this component
as well and add an additional success step for restore in the future.