frontend: remove creatingbackup state
creatingBackup was only used to block the continue button on the
checklist with all warnings regarding the backup.

Simplified by setting all checkboxes to false once continue is
clicked so that the continue button is disabled after clicking,
the view is anyway taken over by either confirm date or insert
sdcard prompt.