frontend: pass `deviceIDs` & `hasAccount` from DeviceSwitch to BB02Settings
to fulfil the tab's conditional (tab is used in the newly created
device settings page) we need these two props.

An alternative to this, would be to omit these two props and
just render all tabs instead of doing conditional rendering
like what we're doing now.

When a page data doesn't exist (example: manage accounts),
instead of not rendering the tab link, we still render it but
in the manage accounts page we can have a message to the
user that "no account is found" or similar.