frontend: fix recursion when device is unplugged during setpassword
When device is unplugged during set password (for example when
password has less than 4 characters warning) the frontend calls
set-password api over and over again.

This was also an issue on master, but it stopped after about 20
calls or so. Seems that it stopped at some point after unmount.

Added useMountedRef hook to only call the api if the device is
mounted.