backend/handlers: skip 0txs to mitigate address poisoning
Address Poisoning attack consists of an attacker creating empty
(0 amount) txs to/from vanity addresses, hoping that the possible
victim will copy/paste the wrong address from the transaction history
when sending/receiving coins. This can happen in some ERC20 tokens,
depending on the implementation of the `transfer` function inside the
token contract (see
https://mirror.xyz/x-explore.eth/cL3d_CyNujXq8XY7ueP4omNXx_IY1EG5Dz0FD0vJ90M
for details).

This hide 0 amount erc20 txs from the `transactions` endpoint, to
mitigate the effects of this attack.