frontend: fix buy redirection from BuyReceiveCTA for btc account
There is a bug if we have multiple btc accounts activated and
only btc accounts, the Buy button in the BuyReceiveCTA component
will redirect the user to the 1st btc account buy
page and this is incorrect. This commit fixes this, and now it'll
redirect the user to the buy account selector page instead.