backend/accounts: fix scanning of first 5 accounts
The step to look for more accounts was only invoked when the previous
account was used. If one had e.g. account 4 unused, but account 5
used, this could prevent discovery of account 5.

It worked anyway in some cases due to a race condition, as
maybeAddHiddenUnusedAccounts could be invoked after account 4
was scanned (e.g. a different coin's account being added after acount
4 was scanned).

tstCheckAccountUsed is added to control the execution of the
checkACcountUsed() goroutine in unit tests. For now we disable it in
unit tests or various unit tests would fail that assume the initial
accounts added are the standard accounts, without the additionally
scanned accounts which are added with a non-deterministic delay. In
the future we can use this callback to add fine-grained unit tests
about accounts scanning.