frontend: change background tertiary in darkmode
The background color of unchecked toggles were hard to see in dark
mode, background tertiary is also used for secondary buttons.

Changed background tertiary in darkmode to fix the toggle color,
this also changes button secondary color.

Another issue is that button secondary and button transparent are
visually very similar, but the transparent style is not a text only
transparent button as the name implies. This can be improved in a
subsequent change.