frontend: support classname prop with toggle component
Toggle props were just passed to the input element, this mostly
makes sense when used as form element. But className does not
make sense on the input element directly it should be the parent
element, as the input is hidden and only used as CSS state.

Changed to add className on the toggle root element so it can
be used for visual styling.

In the codebase className on toggle is only used twice, in last
commit and in togglefirmwarehash component. Last commit expected
that className works properly.

Removed text-small class in togglefirmwarehash as there is no text
and this class does no styling on this component.