frontend: keep accountsync and syncdone apis simple
Api for accountsync and syncdone endpoints contained some logic
to filter for specific code, this was useful in certain situation,
but made generic subscription to all events regardless of code
impossible.

This change removes the code arguemnt and condition to only
subscribe to one code at the time, so it is possible to listen
to all subscriptions.