frontend: using `word-break: break-word` for sidebarLabel
To break word (account name) appropriately, we need to use the
break-word value for word-break css property for sidebarLabel.

Previously, the sidebarLabel is able to break world "appropriately"
when account name is either short or long with spaces. Long names
without spaces will cause the coin logo to shift to the left, which
isn't ideal.