frontend: change from react-scripts to vite
react-script is not recommended anymore, vitejs is activelly
developed and uses modern esbuild under the hood (instead of
webpack).

This commit is the first step of migrating the frontend build to
vite, make buildweb will install all dependencies and make webdev
should work with this change. Make webtest and weblint should be
supported in next commits.