frontends/android: add an authentication feature
Watch-only can benefit from a feature that allows the user to authenticate
when opening the bitboxapp, to protect the privacy. This commit creates
an Auth component in the frontend that can trigger an environment-based
authentication.

On Qt and Webdev the authentication succeds automatically, while on Android
it requires the user to complete the system biometric authentication
every time the app is resumed (.i.e. at startup and when the app comes to
foreground after being put in the background).

The authentication can be requested by the frontend using the 'auth' endpoint.

Authentication messages from backend to frontend are sent via a new 'auth'
notifications, which contains a Typ string:
- 'auth-required': which triggers a new `auth` call from the frontend
- 'auth-ok': which means authentication completed successfully
- 'auth-err': which means authentication failed or has been cancelled
by the user.