backend: do not re-initialize watch-only accounts
When disconnecting a keystore, previously all accounts were removed,
because we had no watch-only accounts.

When registering, all keystore accounts were (re-)loaded, because they
were the only ones to be shown.

With watch-only, we don't want to reinit accounts that are being
watched. This allows a user to connect a keystore e.g. in the middle
of an action inside an account, like displaying an address, without
being disrupted.

The force flag is needed to reinitialize also watch-only accounts if
e.g. the activeTokens field changes for an ETH account. Ideally we can
always dynamically fix the loaded accounts to match the persisted
config, but that is for the future.