backend/pocket: defer CanSignMessage() check
In watch-only accounts the keystore is not connected, so we cannot
decide upfront if the Buy button should be shown. Instead, we check it
when we actually need to sign a message. Currently this can only
happen for the software (testing) keystore and in testnet.

We don't put in the effort to make an error code for error message to
translate it - the BB02 can sign, we don't put in extra effort for
BB01.