frontend/account/info: defer CanVerifyExtendedPublicKey check
In watch-only, we don't have access to the keystore, so we cannot do
this check upfront. Instead, we simply show the button and report an
error if the keystore is unable to verify the xpub.

We don't put in the effort to make error codes for the two error
messages to transalte them:
- "An account must be BTC based to support xpub verification." - can't
happen as the frontend checks for BTC
- "This device/keystore does not support verifying xpubs." - BB02 can
  do it, we don't put in extra effort for BB01. The software keystore is
  only for testing and users don't see it.