frontend: split bitbox02 into settings and wizard
Navigating to 'Settings -> Manage device' shows the device
settings. The BitBox02 component also included all of the
"wizard" (fullscreen) functionality like the unlock animation,
pairing, creating/restoring, etc.

We split the wizard into a new file/component. We mount that in
App.tsx globally, so it can be displayed at any time, not just when
the user is in the `/settings/device-settings/<deviceID>` route. This
is important in watch-only, where the user can connect the BitBox02
when prompted for it (e.g. when trying to verify an address or sign a
transaction) without being removed from whatever worfklow the user is
in.

I copied bitbox02.tsx to wizard.tsx and removed everything
non-essential from the original bitbox02.tsx. In wizard.tsx, almost
everything is the same as in the original, except that the settings
are not rendered and we don't automatically route anywhere after
unlock.

The waitDialog state was removed as it was unused.

--- Routing ---

With watch-only, inserting and unplugging a BitBox02 does not
influence routing anymore. Connecting can happen anywhere (e.g. when
trying to sign).

Instead, we route to the portfolio page when the app goes from no
accounts to some accounts. This happens either at launch when
watch-only account are loaded, or when accounts are loaded for the
first time when unlocking a keystore.

For the bitbox01 and the bitbox02 bootloader, we keep routing to the
settings, the bb01 unlock screen and bb02 bootloader screen are not
full screen and are not mounted globally.
This means that
- if a bb01 user needs to insert their device when signing a
  transaction, they will be removed from that screen. We don't put in
  extra effort into the bb01
- if a bb02 user needs to insert but ends up having to upgrade the fw,
  they will also be removed from their current workflow. This should
  be a very rare, ideally never.