backend/bitsurance: update bitsurance check to fetch account details
Previous implementation was only checking if the insurance status was
active and saving in accounts.json the bitsurance id for insured
accounts.

With this, the insurance status is persisted in the accounts.json, instead
of the bitsurance id, and all the insurance details are fetched from
the server. This will be used in the next commit to display the insurance
details in a new dashboard page.