backend/handlers: fix loading ERC20 tokens
Regression from 18aea1dea, which started loading the persisted
accounts for all accounts, not just ETH. For ERC20 token accounts
there is no corresponding persisted account, so the token would fail
to be shown in the frontend.

We can just use  `account.Config().Config` instead, which is loaded
from the persisted config, and which has meaningful data for tokens -
e.g. the same signing config as the parent ETH account.