frontend: fix flickering dialog on close
When closing dialog (either by pressing `esc`,  or clicking the update /
close button), sometimes the dialog flickers. This only happens on desktop.
The reason for this is because we previously fire `deactivateModal()`
inside a timeout (300ms) to await for transitions to finish.

This is working fine on mobile, but not on desktop. On desktop we
should call the `deactivateModal` with no delay to prevent
unnecessary re-render, and only wait for animation (css transition)
when there is one (such as on mobile).

To achieve this, we're using `transitionend` event listener to
more accurately wait for the transition to end and fire
`deactivateModal` or not wait at all when there's no transition,
which ultimately fixes the flickering issue.