backend/accounts: properly mark discovered accounts as watchonly
Before this commit, accounts that are added for scanning in the
background (HiddenBecauseUnused) would all be marked watchonly, even
though they are not shown to the user yet (the account is only shown
if the user manually adds it or if it is found to be non-empty
automatically). This way, the account would appear automatically once
it received funds, even if it was not visible before and the keystore
is never connected again.

Instead, it should be marked watchonly when the account is shown to
the user and when the keystore is connected, as all other accounts.