frontend: show actual account name in sidebar
The sidebar header used to display ACCOUNTS (uppercase) but with
watch-only this now includes the wallet name (user input).

Removed uppercase styling so that the BitBoxApp shows the exact
name that the user has choosen. The user can decide to use a name
that is all uppercase. The app should just display the name as
choosen and not transform to uppercase.