frontend/sidebar: add fingerprint to accounts group to disambiguate
If multiple keystores have the same name (commonly when you add
keystores from one seed using multiple different passphrases), we
append the fingerprint to disambiguate the groups.

The app does not know which keystores are passphrase accounts and
which are not, so we use the name as the closest thing we have to
disambiguate.