frontend: fix centered badge icon
The icon in the badge was not perfectly centered. Reason is
that the icon has an uneven width (9px) so it rounds the
position to be 1px off.

Changed so that the icon is now a max of 10px wide. Ideally
all icons in the future that are used in badges should have
an even width.